moves all interfaces to own file: - Interfaces that correspond to types from backend in own file - Custom frontend types in interfaces.ts file
